# docs/app/swagger.yaml
basePath: /api/v1
definitions:
...
  placio-app_ent.Business:
    properties:
      business_settings:
        additionalProperties: true
        description: BusinessSettings holds the value of the "business_settings" field.
        type: object
      cover_image:
        description: CoverImage holds the value of the "cover_image" field.
        type: string
      description:
        description: Description holds the value of the "description" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/placio-app_ent.BusinessEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the BusinessQuery when eager-loading is set.
      email:
        description: Email holds the value of the "email" field.
        type: string
      id:
        description: ID of the ent.
        type: string
      latitude:
        description: Latitude holds the value of the "latitude" field.
        type: string
      location:
        description: Location holds the value of the "location" field.
        type: string
      longitude:
        description: Longitude holds the value of the "longitude" field.
        type: string
      map_coordinates:
        additionalProperties: true
        description: MapCoordinates holds the value of the "map_coordinates" field.
        type: object
      name:
        description: Name holds the value of the "name" field.
        type: string
      phone:
        description: Phone holds the value of the "phone" field.
        type: string
      picture:
        description: Picture holds the value of the "picture" field.
        type: string
      relevance_score:
        description: RelevanceScore holds the value of the "relevance_score" field.
        type: number
      search_text:
        description: SearchText holds the value of the "search_text" field.
        type: string
      url:
        description: URL holds the value of the "url" field.
        type: string
      website:
        description: Website holds the value of the "website" field.
        type: string
    type: object
  placio-app_ent.BusinessEdges:
    properties:
      business_account_settings:
        allOf:
        - $ref: '#/definitions/placio-app_ent.AccountSettings'
        description: BusinessAccountSettings holds the value of the business_account_settings
          edge.
      businessFollowEvents:
        description: BusinessFollowEvents holds the value of the businessFollowEvents
          edge.
        items:
          $ref: '#/definitions/placio-app_ent.BusinessFollowEvent'
        type: array
      categories:
        description: Categories holds the value of the categories edge.
        items:
          $ref: '#/definitions/placio-app_ent.Category'
        type: array
      categoryAssignments:
        description: CategoryAssignments holds the value of the categoryAssignments
          edge.
        items:
          $ref: '#/definitions/placio-app_ent.CategoryAssignment'
        type: array
      events:
        description: Events holds the value of the events edge.
        items:
          $ref: '#/definitions/placio-app_ent.Event'
        type: array
      faqs:
        description: Faqs holds the value of the faqs edge.
        items:
          $ref: '#/definitions/placio-app_ent.FAQ'
        type: array
      followedBusinesses:
        description: FollowedBusinesses holds the value of the followedBusinesses
          edge.
        items:
          $ref: '#/definitions/placio-app_ent.BusinessFollowBusiness'
        type: array
      followedUsers:
        description: FollowedUsers holds the value of the followedUsers edge.
        items:
          $ref: '#/definitions/placio-app_ent.BusinessFollowUser'
        type: array
      followerBusinesses:
        description: FollowerBusinesses holds the value of the followerBusinesses
          edge.
        items:
          $ref: '#/definitions/placio-app_ent.BusinessFollowBusiness'
        type: array
      followerUsers:
        description: FollowerUsers holds the value of the followerUsers edge.
        items:
          $ref: '#/definitions/placio-app_ent.UserFollowBusiness'
        type: array
      places:
        description: Places holds the value of the places edge.
        items:
          $ref: '#/definitions/placio-app_ent.Place'
        type: array
      posts:
        description: Posts holds the value of the posts edge.
        items:
          $ref: '#/definitions/placio-app_ent.Post'
        type: array
      ratings:
        description: Ratings holds the value of the ratings edge.
        items:
          $ref: '#/definitions/placio-app_ent.Rating'
        type: array
      userBusinesses:
        description: UserBusinesses holds the value of the userBusinesses edge.
        items:
          $ref: '#/definitions/placio-app_ent.UserBusiness'
        type: array
    type: object
...
  placio-app_ent.Comment:
    properties:
      Content:
        description: Content holds the value of the "Content" field.
        type: string
      CreatedAt:
        description: CreatedAt holds the value of the "CreatedAt" field.
        type: string
      UpdatedAt:
        description: UpdatedAt holds the value of the "UpdatedAt" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/placio-app_ent.CommentEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the CommentQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: string
    type: object
  placio-app_ent.CommentEdges:
    properties:
      post:
        allOf:
        - $ref: '#/definitions/placio-app_ent.Post'
        description: Post holds the value of the post edge.
      user:
        allOf:
        - $ref: '#/definitions/placio-app_ent.User'
        description: User holds the value of the user edge.
    type: object
  placio-app_ent.Event:
    properties:
      EventType:
        allOf:
        - $ref: '#/definitions/placio-app_ent_event.EventType'
        description: EventType holds the value of the "EventType" field.
      cover_image:
        description: CoverImage holds the value of the "cover_image" field.
        type: string
      createdAt:
        description: CreatedAt holds the value of the "createdAt" field.
        type: string
      description:
        description: Description holds the value of the "description" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/placio-app_ent.EventEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the EventQuery when eager-loading is set.
      end_date:
        description: EndDate holds the value of the "end_date" field.
        type: string
      end_time:
        description: EndTime holds the value of the "end_time" field.
        type: string
      event_settings:
        additionalProperties: true
        description: EventSettings holds the value of the "event_settings" field.
        type: object
      frequency:
        allOf:
        - $ref: '#/definitions/placio-app_ent_event.Frequency'
        description: Frequency holds the value of the "frequency" field.
      frequency_day_of_month:
        description: FrequencyDayOfMonth holds the value of the "frequency_day_of_month"
          field.
        type: string
      frequency_day_of_week:
        description: FrequencyDayOfWeek holds the value of the "frequency_day_of_week"
          field.
        type: string
      frequency_interval:
        description: FrequencyInterval holds the value of the "frequency_interval"
          field.
        type: string
      frequency_month_of_year:
        description: FrequencyMonthOfYear holds the value of the "frequency_month_of_year"
          field.
        type: string
      id:
        description: ID of the ent.
        type: string
      latitude:
        description: Latitude holds the value of the "latitude" field.
        type: string
      location:
        description: Location holds the value of the "location" field.
        type: string
      longitude:
        description: Longitude holds the value of the "longitude" field.
        type: string
      map_coordinates:
        additionalProperties: true
        description: MapCoordinates holds the value of the "map_coordinates" field.
        type: object
      name:
        description: Name holds the value of the "name" field.
        type: string
      relevance_score:
        description: RelevanceScore holds the value of the "relevance_score" field.
        type: number
      search_text:
        description: SearchText holds the value of the "search_text" field.
        type: string
      start_date:
        description: StartDate holds the value of the "start_date" field.
        type: string
      start_time:
        description: StartTime holds the value of the "start_time" field.
        type: string
      status:
        description: Status holds the value of the "status" field.
        type: string
      tags:
        description: Tags holds the value of the "tags" field.
        type: string
      time_zone:
        description: TimeZone holds the value of the "time_zone" field.
        type: string
      title:
        description: Title holds the value of the "title" field.
        type: string
      updatedAt:
        description: UpdatedAt holds the value of the "updatedAt" field.
        type: string
      url:
        description: URL holds the value of the "url" field.
        type: string
      venue_address:
        description: VenueAddress holds the value of the "venue_address" field.
        type: string
      venue_city:
        description: VenueCity holds the value of the "venue_city" field.
        type: string
      venue_country:
        description: VenueCountry holds the value of the "venue_country" field.
        type: string
      venue_email:
        description: VenueEmail holds the value of the "venue_email" field.
        type: string
      venue_lat:
        description: VenueLat holds the value of the "venue_lat" field.
        type: string
      venue_lon:
        description: VenueLon holds the value of the "venue_lon" field.
        type: string
      venue_name:
        description: VenueName holds the value of the "venue_name" field.
        type: string
      venue_phone:
        description: VenuePhone holds the value of the "venue_phone" field.
        type: string
      venue_state:
        description: VenueState holds the value of the "venue_state" field.
        type: string
      venue_type:
        allOf:
        - $ref: '#/definitions/placio-app_ent_event.VenueType'
        description: VenueType holds the value of the "venue_type" field.
      venue_url:
        description: VenueURL holds the value of the "venue_url" field.
        type: string
      venue_zip:
        description: VenueZip holds the value of the "venue_zip" field.
        type: string
    type: object
...
  placio-app_ent.Like:
    properties:
      createdAt:
        description: CreatedAt holds the value of the "createdAt" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/placio-app_ent.LikeEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the LikeQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: string
      like:
        description: True for like, False for dislike
        type: boolean
      updatedAt:
        description: UpdatedAt holds the value of the "updatedAt" field.
        type: string
    type: object
...
  placio-app_ent.Media:
    properties:
      CreatedAt:
        description: CreatedAt holds the value of the "CreatedAt" field.
        type: string
      MediaType:
        description: image, gif, or video
        type: string
      URL:
        description: URL holds the value of the "URL" field.
        type: string
      UpdatedAt:
        description: UpdatedAt holds the value of the "UpdatedAt" field.
        type: string
      dislikeCount:
        description: Count of dislikes for this media.
        type: integer
      edges:
        allOf:
        - $ref: '#/definitions/placio-app_ent.MediaEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the MediaQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: string
      likeCount:
        description: Count of likes for this media.
        type: integer
    type: object
...
  placio-app_ent.Place:
    properties:
      additional_info:
        additionalProperties: true
        description: AdditionalInfo holds the value of the "additional_info" field.
        type: object
      availability:
        additionalProperties: true
        description: Availability holds the value of the "availability" field.
        type: object
      city:
        description: City holds the value of the "city" field.
        type: string
      country:
        description: Country holds the value of the "country" field.
        type: string
      cover_image:
        description: CoverImage holds the value of the "cover_image" field.
        type: string
      description:
        description: Description holds the value of the "description" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/placio-app_ent.PlaceEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the PlaceQuery when eager-loading is set.
      email:
        description: Email holds the value of the "email" field.
        type: string
      features:
        description: Features holds the value of the "features" field.
        items:
          type: string
        type: array
      id:
        description: ID of the ent.
        type: string
      images:
        description: Images holds the value of the "images" field.
        items:
          type: string
        type: array
      latitude:
        description: Latitude holds the value of the "latitude" field.
        type: string
      location:
        description: Location holds the value of the "location" field.
        type: string
      longitude:
        description: Longitude holds the value of the "longitude" field.
        type: string
      map_coordinates:
        additionalProperties: true
        description: MapCoordinates holds the value of the "map_coordinates" field.
        type: object
      name:
        description: Name holds the value of the "name" field.
        type: string
      opening_hours:
        additionalProperties: true
        description: OpeningHours holds the value of the "opening_hours" field.
        type: object
      payment_options:
        additionalProperties: true
        description: PaymentOptions holds the value of the "payment_options" field.
        type: object
      phone:
        description: Phone holds the value of the "phone" field.
        type: string
      picture:
        description: Picture holds the value of the "picture" field.
        type: string
      place_settings:
        additionalProperties: true
        description: PlaceSettings holds the value of the "place_settings" field.
        type: object
      relevance_score:
        description: RelevanceScore holds the value of the "relevance_score" field.
        type: number
      search_text:
        description: SearchText holds the value of the "search_text" field.
        type: string
      social_media:
        additionalProperties: true
        description: SocialMedia holds the value of the "social_media" field.
        type: object
      special_offers:
        description: SpecialOffers holds the value of the "special_offers" field.
        type: string
      state:
        description: State holds the value of the "state" field.
        type: string
      sustainability_score:
        description: SustainabilityScore holds the value of the "sustainability_score"
          field.
        type: number
      tags:
        description: Tags holds the value of the "tags" field.
        items:
          type: string
        type: array
      type:
        description: Type holds the value of the "type" field.
        type: string
      website:
        description: Website holds the value of the "website" field.
        type: string
    type: object
...
  placio-app_ent.Review:
    properties:
      content:
        description: User's review to the business/place/event.
        type: string
      createdAt:
        description: When the review was created.
        type: string
      dislikeCount:
        description: Count of dislikes for this review.
        type: integer
      edges:
        allOf:
        - $ref: '#/definitions/placio-app_ent.ReviewEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ReviewQuery when eager-loading is set.
      flag:
        description: Flag for this review.
        type: string
      id:
        description: ID of the ent.
        type: string
      likeCount:
        description: Count of likes for this review.
        type: integer
      score:
        description: Score should be between 1 and 5.
        type: number
    type: object
  placio-app_ent.ReviewEdges:
    properties:
      business:
        allOf:
        - $ref: '#/definitions/placio-app_ent.Business'
        description: The business that was reviewed.
      comments:
        description: The comments related to the review.
        items:
          $ref: '#/definitions/placio-app_ent.Comment'
        type: array
      event:
        allOf:
        - $ref: '#/definitions/placio-app_ent.Event'
        description: The event that was reviewed.
      likes:
        description: The likes related to the review.
        items:
          $ref: '#/definitions/placio-app_ent.Like'
        type: array
      medias:
        description: The media content related to the review.
        items:
          $ref: '#/definitions/placio-app_ent.Media'
        type: array
      place:
        allOf:
        - $ref: '#/definitions/placio-app_ent.Place'
        description: The place that was reviewed.
      user:
        allOf:
        - $ref: '#/definitions/placio-app_ent.User'
        description: The user who wrote the review.
    type: object
...
  placio-app_ent.User:
    properties:
      app_settings:
        additionalProperties: true
        description: AppSettings holds the value of the "app_settings" field.
        type: object
      auth0_data:
        allOf:
        - $ref: '#/definitions/management.User'
        description: Auth0Data holds the value of the "auth0_data" field.
      auth0_id:
        description: Auth0ID holds the value of the "auth0_id" field.
        type: string
      bio:
        description: Bio holds the value of the "bio" field.
        type: string
      cover_image:
        description: CoverImage holds the value of the "cover_image" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/placio-app_ent.UserEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the UserQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: string
      latitude:
        description: Latitude holds the value of the "latitude" field.
        type: string
      location:
        description: Location holds the value of the "location" field.
        type: string
      longitude:
        description: Longitude holds the value of the "longitude" field.
        type: string
      map_coordinates:
        additionalProperties: true
        description: MapCoordinates holds the value of the "map_coordinates" field.
        type: object
      name:
        description: Name holds the value of the "name" field.
        type: string
      picture:
        description: Picture holds the value of the "picture" field.
        type: string
      relevance_score:
        description: RelevanceScore holds the value of the "relevance_score" field.
        type: number
      search_text:
        description: SearchText holds the value of the "search_text" field.
        type: string
      user_settings:
        additionalProperties: true
        description: UserSettings holds the value of the "user_settings" field.
        type: object
      username:
        description: Username holds the value of the "username" field.
        type: string
      website:
        description: Website holds the value of the "website" field.
        type: string
    type: object
...
  placio-app_ent_event.VenueType:
    enum:
    - online
    - in_person
    - hybrid
    type: string
    x-enum-varnames:
    - VenueTypeOnline
    - VenueTypeInPerson
    - VenueTypeHybrid
...
host: https://api.palnight.com
info:
  contact:
    email: support@placio.io
    name: Placio Ltd
    url: https://placio.io
  description: This is the documentation for the Placio Application Api
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://placio.io/terms
  title: Placio Application Api
  version: "0.01"
paths:
...
  /reviews/:
    get:
      consumes:
      - application/json
      description: Get reviews based on query parameters
      produces:
      - application/json
      responses:
        "200":
          description: Array of Reviews
          schema:
            items:
              $ref: '#/definitions/placio-app_ent.Review'
            type: array
        "500":
          description: Error message
          schema:
            type: string
      summary: Get reviews by query
      tags:
      - Review
    post:
      consumes:
      - application/json
      description: Rate a place, event, or business
      parameters:
      - description: Item Type (place, event, business)
        in: query
        name: type
        required: true
        type: string
      - description: Item ID (placeID, eventID, businessID)
        in: query
        name: id
        required: true
        type: string
      - description: Rating Score
        in: body
        name: score
        required: true
        schema:
          type: number
      - description: Review Content
        in: body
        name: content
        required: true
        schema:
          type: string
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully rated [itemType]
          schema:
            type: string
        "400":
          description: Invalid item type
          schema:
            type: string
        "500":
          description: Error message
          schema:
            type: string
      summary: Review a Place, Event, or Business
      tags:
      - Review
  /reviews/{reviewID}:
    delete:
      consumes:
      - application/json
      description: Remove a review based on reviewID
      parameters:
      - description: Review ID
        in: path
        name: reviewID
        required: true
        type: string
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted review
          schema:
            type: string
        "500":
          description: Error message
          schema:
            type: string
      summary: Remove a review
      tags:
      - Review
    get:
      consumes:
      - application/json
      description: Retrieve a review using its ID
      parameters:
      - description: Review ID
        in: path
        name: reviewID
        required: true
        type: string
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Review data
          schema:
            $ref: '#/definitions/placio-app_ent.Review'
        "500":
          description: Error message
          schema:
            type: string
      summary: Get review by ID
      tags:
      - Review
    put:
      consumes:
      - application/json
      description: Update the content of a review based on reviewID
      parameters:
      - description: Review ID
        in: path
        name: reviewID
        required: true
        type: string
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: New Review Content
        in: body
        name: content
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated review
          schema:
            type: string
        "500":
          description: Error message
          schema:
            type: string
      summary: Update review content
      tags:
      - Review
  /reviews/{reviewID}/addMedia:
    post:
      consumes:
      - application/json
      description: Add media to a review based on reviewID
      parameters:
      - description: Review ID
        in: path
        name: reviewID
        required: true
        type: string
      - description: Media File
        in: body
        name: string
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully added media to review
          schema:
            type: string
        "500":
          description: Error message
          schema:
            type: string
      summary: Add media to a review
      tags:
      - Review
  /reviews/{reviewID}/dislike:
    post:
      consumes:
      - application/json
      description: Dislike a review based on reviewID
      parameters:
      - description: Review ID
        in: path
        name: reviewID
        required: true
        type: string
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully disliked review
          schema:
            type: string
        "500":
          description: Error message
          schema:
            type: string
      summary: Dislike a review
      tags:
      - Review
  /reviews/{reviewID}/like:
    post:
      consumes:
      - application/json
      description: Like a review based on reviewID
      parameters:
      - description: Review ID
        in: path
        name: reviewID
        required: true
        type: string
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully liked review
          schema:
            type: string
        "500":
          description: Error message
          schema:
            type: string
      summary: Like a review
      tags:
      - Review
...
schemes:
- http
- https
swagger: "2.0"